+++ src/backend/profile_types.ts
@@ -11,7 +11,7 @@ export interface CallFrame {
 export interface DmaRecord {
 	reg?: number; // & 0x1000 => CPU
 	dat?: number; // 16, 32 or 64 bit (lower 32 bit)
-	datHi?: number; // 16, 32 or 64 bit (upper 32 bit)
+	datHi?: number; // 64 bit (upper 32 bit)
 	size?: number; // 2 = 16 bit, 4 = 32 bit, 8 = 64 bit
 	addr?: number;
 	evt?: number;